examples/cv-model/requirements.txt
tensorflow==2.4.0
numpy==1.19.5
Pillow==8.2.0
opencv-python==4.5.1.48
Flask==2.0.1
gunicorn==20.1.0
prometheus-client==0.11.0
python-dotenv==0.19.0
pydantic==1.8.2
requests==2.26.0 